Syntax errors now include stderr text
* AzCommandSyntaxError exception now includes the original stderr text
* Replaced elif block with dictionary of lambdas
