+++ src/azwrap/__init__.py
@@ -1,1 +1,2 @@
 from azwrap.core import Az
+from azwrap.exceptions import AzCommandSyntaxError, AzParserError, AzRequestError, AzResourceNotFoundError, AzUnsupportedCommand
